CRITICAL FIX: Remove dangerous uint to ushort cast in scene entity activation - MaxSceneEntities is configurable!

+++ MonoGame/Atomic.Net.MonoGame.Benchmarks/Core/SparseArrayBenchmark.cs
@@ -19,14 +19,14 @@ public class SparseArrayBenchmark
     public void Setup()
     {
         _array = new int?[Capacity];
-        _sparseArray = new SparseArray<int>((ushort)Capacity);
+        _sparseArray = new SparseArray<int>((uint)Capacity);
 
-        for (ushort i = 0; i < Capacity; i++)
+        for (uint i = 0; i < Capacity; i++)
         {
             if (_rng.NextDouble() < FillPercent)
             {
-                _array[i] = i;
-                _sparseArray.Set(i, i);
+                _array[i] = (int)i;
+                _sparseArray.Set(i, (int)i);
             }
         }
     }
