+++ MonoGame/Atomic.Net.MonoGame.Benchmarks/Core/Units/BitwiseAndBenchmark.cs
@@ -95,11 +95,11 @@ public class BitwiseAndBenchmark
             // Only set in SparseArray if true (sparse storage)
             if (leftValue)
             {
-                _leftSparse.Set((ushort)i, true);
+                _leftSparse.Set((uint)i, true);
             }
             if (rightValue)
             {
-                _rightSparse.Set((ushort)i, true);
+                _rightSparse.Set((uint)i, true);
             }
         }
